backend/requirements.txt
appdirs==1.4.4
ariadne==0.13.0
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.5
django-js-asset==1.2.2
djangorestframework==3.12.4
graphql-core==3.1.5
isort==5.9.2
mypy-extensions==0.4.3
pathspec==0.9.0
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
starlette==0.14.2
tomli==1.0.4
typing-extensions==3.10.0.0
